+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.13.0] - 2026-03-27
+
+## Changed
+
+- Text search now runs in parallel for large documents in
+[app.rs](src/ui/app.rs), improving performance.
+
 ## [0.12.0] - 2026-02-05
 
 ## Added